refactor OwsExceptionReport, and deps...
