+++ core/src/main/java/org/n52/iceland/binding/json/JSONBinding.java
@@ -26,6 +26,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.n52.iceland.binding.Binding;
 import org.n52.iceland.binding.SimpleBinding;
 import org.n52.iceland.coding.OperationKey;
+import org.n52.iceland.coding.decode.Decoder;
 import org.n52.iceland.coding.decode.OperationDecoderKey;
 import org.n52.iceland.exception.HTTPException;
 import org.n52.iceland.exception.ows.NoApplicableCodeException;
@@ -48,11 +49,6 @@ import org.n52.iceland.binding.PathBindingKey;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.ImmutableSet;
-import java.util.logging.Level;
-import org.n52.iceland.coding.decode.ConformanceClassDecoder;
-import org.n52.iceland.coding.decode.Decoder;
-import org.n52.iceland.exception.CodingException;
-import org.n52.iceland.exception.UnsupportedDecoderInputException;
 
 /**
  * {@link Binding} implementation for JSON encoded requests
@@ -136,12 +132,7 @@ public class JSONBinding extends SimpleBinding {
             if (decoder == null) {
                 throw new NoDecoderForKeyException(key);
             }
-            AbstractServiceRequest<?> sosRequest;
-            try {
-                sosRequest = decoder.decode(json);
-            } catch (CodingException | UnsupportedDecoderInputException ex) {
-                throw new NoApplicableCodeException().causedBy(ex);
-            }
+            AbstractServiceRequest<?> sosRequest = decoder.decode(json);
             sosRequest.setRequestContext(getRequestContext(request));
             return sosRequest;
         } catch (IOException ioe) {
