+++ core/src/main/java/org/n52/iceland/binding/kvp/KvpBinding.java
@@ -34,6 +34,7 @@ import org.n52.iceland.binding.MediaTypeBindingKey;
 import org.n52.iceland.binding.PathBindingKey;
 import org.n52.iceland.binding.SimpleBinding;
 import org.n52.iceland.coding.OperationKey;
+import org.n52.iceland.coding.decode.Decoder;
 import org.n52.iceland.coding.decode.DecoderKey;
 import org.n52.iceland.coding.decode.OperationDecoderKey;
 import org.n52.iceland.config.annotation.Configurable;
@@ -44,8 +45,8 @@ import org.n52.iceland.exception.ows.concrete.InvalidServiceParameterException;
 import org.n52.iceland.exception.ows.concrete.MissingRequestParameterException;
 import org.n52.iceland.exception.ows.concrete.NoDecoderForKeyException;
 import org.n52.iceland.exception.ows.concrete.VersionNotSupportedException;
-import org.n52.iceland.ogc.ows.OWSConstants;
-import org.n52.iceland.ogc.ows.OWSConstants.RequestParams;
+import org.n52.iceland.ogc.ows.ExtendedOWSConstants;
+import org.n52.iceland.ogc.ows.ExtendedOWSConstants.RequestParams;
 import org.n52.iceland.ogc.sos.ConformanceClasses;
 import org.n52.iceland.ogc.sos.Sos2Constants;
 import org.n52.iceland.ogc.sos.SosConstants;
@@ -58,12 +59,6 @@ import org.n52.iceland.util.http.MediaTypes;
 
 import com.google.common.collect.ImmutableSet;
 import java.util.Arrays;
-import java.util.logging.Level;
-import org.n52.iceland.coding.decode.ConformanceClassDecoder;
-import org.n52.iceland.coding.decode.Decoder;
-import org.n52.iceland.exception.CodingException;
-import org.n52.iceland.exception.UnsupportedDecoderInputException;
-import org.n52.iceland.exception.ows.NoApplicableCodeException;
 
 /**
  * OWS binding for Key-Value-Pair (HTTP-Get) requests
@@ -167,7 +162,7 @@ public class KvpBinding extends SimpleBinding {
     }
 
     protected boolean isGetCapabilities(Map<String, String> map) throws OwsExceptionReport {
-        return OWSConstants.Operations.GetCapabilities.name().equals(getRequestParameterValue(map));
+        return ExtendedOWSConstants.Operations.GetCapabilities.name().equals(getRequestParameterValue(map));
     }
 
     public String getRequestParameterValue(Map<String, String> map) throws OwsExceptionReport {
@@ -195,12 +190,7 @@ public class KvpBinding extends SimpleBinding {
         Decoder<AbstractServiceRequest<?>, Map<String, String>> decoder = getDecoder(k);
         LOGGER.trace("Using {} to decode paramers: {}", decoder, Arrays.toString(parameterValueMap.entrySet().toArray()));
         if (decoder != null) {
-            AbstractServiceRequest<?> request;
-            try {
-                request = decoder.decode(parameterValueMap);
-            } catch (CodingException | UnsupportedDecoderInputException ex) {
-                throw new NoApplicableCodeException().causedBy(ex);
-            }
+            AbstractServiceRequest<?> request = decoder.decode(parameterValueMap);
             if (includeOriginal) {
                 request.setOriginalRequest(urlJoiner.join(req.getRequestURL(), req.getQueryString()));
             }
