+++ core/src/main/java/org/n52/iceland/exception/ows/concrete/VersionNotSupportedException.java
@@ -17,7 +17,7 @@
 package org.n52.iceland.exception.ows.concrete;
 
 import org.n52.iceland.exception.ows.InvalidParameterValueException;
-import org.n52.iceland.ogc.ows.OWSConstants.RequestParams;
+import org.n52.iceland.ogc.ows.ExtendedOWSConstants.RequestParams;
 
 /**
  * @author <a href="mailto:c.autermann@52north.org">Christian Autermann</a>
