+++ core/src/main/java/org/n52/iceland/ogc/ows/ExtendedOWSConstants.java
@@ -16,9 +16,7 @@
  */
 package org.n52.iceland.ogc.ows;
 
-import javax.xml.namespace.QName;
 import org.n52.iceland.exception.ows.InvalidParameterValueException;
-import org.n52.iceland.exception.ows.OwsExceptionCode;
 import org.n52.iceland.w3c.SchemaLocation;
 
 /**
@@ -26,76 +24,15 @@ import org.n52.iceland.w3c.SchemaLocation;
  *
  * @since 1.0.0
  */
-public interface OWSConstants {
-
-    //fixme: add version...
-    String NS_OWS = "http://www.opengis.net/ows/1.1";
-
-    String NS_OWS_PREFIX = "ows";
-
-    String SCHEMA_LOCATION_URL_OWS = "http://schemas.opengis.net/ows/1.1.0/owsAll.xsd";
-
-    String SCHEMA_LOCATION_URL_OWS_EXCEPTIONREPORT = "http://schemas.opengis.net/ows/1.1.0/owsExceptionReport.xsd";
+public interface ExtendedOWSConstants extends OWSConstants  {
 
     SchemaLocation OWS_110_SCHEMA_LOCATION = new SchemaLocation(NS_OWS, SCHEMA_LOCATION_URL_OWS);
 
     SchemaLocation OWS_110_EXCEPTION_REPORT_SCHEMA_LOCATION = new SchemaLocation(NS_OWS_PREFIX,
             SCHEMA_LOCATION_URL_OWS_EXCEPTIONREPORT);
 
-    String NS_OWS_100 = "http://www.opengis.net/ows/1.0";
-
-    String SCHEMA_LOCATION_URL_OWS100 = "http://schemas.opengis.net/ows/1.0.0/owsAll.xsd";
-
     SchemaLocation OWS_100_SCHEMA_LOCATION = new SchemaLocation(NS_OWS_100, SCHEMA_LOCATION_URL_OWS100);
 
-    // exception messages
-    String SOAP_REASON_INVALID_PARAMETER_VALUE = "The request contained an invalid parameter value.";
-
-    String SOAP_REASON_INVALID_UPDATE_SEQUENCES
-            = "The value of the updateSequence parameter in the GetCapabilities operation request was greater than the current value of the service metadata updateSequence number.";
-
-    String SOAP_REASON_MISSING_PARAMETER_VALUE
-            = "The request did not include a value for a required parameter and this server does not declare a default value for it.";
-
-    String SOAP_REASON_NO_APPLICABLE_CODE = "A server exception was encountered.";
-
-    String SOAP_REASON_NO_DATA_AVAILABLE = "There are no data available.";
-
-    String SOAP_REASON_OPERATION_NOT_SUPPORTED = "The requested operation is not supported by this server.";
-
-    String SOAP_REASON_OPTION_NOT_SUPPORTED =
-            "The request included/targeted an option that is not supported by this server.";
-
-    String SOAP_REASON_REQUEST_EXTENSION_NOT_SUPPORTED =
-            "The request included an extension that is not supported by this server.";
-
-    String SOAP_REASON_VERSION_NEGOTIATION_FAILED =
-            "The list of versions in the ‘AcceptVersions’ parameter value of the GetCapabilities operation request did not include any version supported by this server.";
-
-    String SOAP_REASON_RESPONSE_EXCEEDS_SIZE_LIMIT =
-            "The requested result set exceeds the response size limit of this service and thus cannot be delivered.";
-
-    @Deprecated // SOS-specific
-    String SOAP_REASON_INVALID_PROPERTY_OFFERING_COMBINATION =
-            "Observations for the requested combination of observedProperty and offering do not use SWE Common encoded results.";
-
-    String SOAP_REASON_UNKNOWN = "A server exception was encountered.";
-
-    String EN_EXCEPTION = "Exception";
-
-    String EN_EXCEPTION_CODE = "exceptionCode";
-
-    String EN_LOCATOR = "locator";
-
-    String EN_EXCEPTION_TEXT = "ExceptionText";
-
-    QName QN_EXCEPTION = new QName(NS_OWS, EN_EXCEPTION, NS_OWS_PREFIX);
-
-    QName QN_EXCEPTION_TEXT = new QName(NS_OWS, EN_EXCEPTION_TEXT, NS_OWS_PREFIX);
-
-    QName QN_NO_APPLICABLE_CODE = new QName(NS_OWS, OwsExceptionCode.NoApplicableCode.name(), NS_OWS_PREFIX);
-
-    String VERSION = "1.1.0";
 
     /**
      * enum with parameter names for getCapabilities request
@@ -111,14 +48,6 @@ public interface OWSConstants {
         CapabilitiesId;
 
         public static final String DYNAMIC_CAPABILITIES_IDENTIFIER = "dynamic";
-    }
-
-    /**
-     * the names of the operations supported by all versions of the service
-     * specification
-     */
-    enum Operations {
-        GetCapabilities;
     }
 
     /**
