--- util/src/main/java/org/n52/iceland/exception/CodingException.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2015-2016 52°North Initiative for Geospatial Open Source
- * Software GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.n52.iceland.exception;
-
-/**
- *
- * @author Matthes Rieke <m.rieke@52north.org>
- */
-public class CodingException extends Exception {
-
-    public CodingException(String message) {
-        super(message);
-    }
-
-    public CodingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public CodingException(Throwable cause) {
-        super(cause);
-    }
-
-
-}
