+++ util/src/main/java/org/n52/iceland/exception/ows/NoApplicableCodeException.java
