+++ util/src/main/java/org/n52/iceland/exception/ows/OwsExceptionCode.java
