+++ util/src/main/java/org/n52/iceland/ogc/ows/ExceptionCode.java
